lib/internal/Magento/Framework/App/TemplateTypesInterface was moved to a lib/internal/Magento/Framework/Mail/TemplateTypesInterface namespase
