+++ app/code/Magento/Email/Test/Unit/Model/AbstractTemplateTest.php
@@ -18,8 +18,8 @@ use Magento\Email\Model\Template\FilterFactory;
 use Magento\Email\Model\TemplateFactory;
 use Magento\Framework\App\Area;
 use Magento\Framework\App\Config\ScopeConfigInterface;
-use Magento\Framework\App\TemplateTypesInterface;
 use Magento\Framework\Filesystem;
+use Magento\Framework\Mail\TemplateTypesInterface;
 use Magento\Framework\TestFramework\Unit\Helper\ObjectManager;
 use Magento\Framework\View\Asset\Repository;
 use Magento\Framework\View\DesignInterface;
