+++ app/code/Magento/Email/view/adminhtml/templates/template/edit.phtml
@@ -4,7 +4,7 @@
  * See COPYING.txt for license details.
  */
 
-use Magento\Framework\App\TemplateTypesInterface;
+use Magento\Framework\Mail\TemplateTypesInterface;
 
 // phpcs:disable Generic.Files.LineLength.TooLong
