+++ app/code/Magento/Newsletter/Block/Adminhtml/Template/Grid.php
@@ -12,7 +12,7 @@
 namespace Magento\Newsletter\Block\Adminhtml\Template;
 
 use Magento\Backend\Block\Widget\Grid as WidgetGrid;
-use Magento\Framework\App\TemplateTypesInterface;
+use Magento\Framework\Mail\TemplateTypesInterface;
 
 class Grid extends \Magento\Backend\Block\Widget\Grid\Extended
 {
