+++ app/code/Magento/Newsletter/Controller/Adminhtml/Template/Save.php
@@ -6,7 +6,7 @@
 namespace Magento\Newsletter\Controller\Adminhtml\Template;
 
 use Magento\Framework\App\Action\HttpPostActionInterface as HttpPostActionInterface;
-use Magento\Framework\App\TemplateTypesInterface;
+use Magento\Framework\Mail\TemplateTypesInterface;
 use Magento\Framework\Exception\LocalizedException;
 
 /**
