+++ app/code/Magento/Newsletter/Test/Unit/Model/TemplateTest.php
@@ -12,10 +12,10 @@ use Magento\Email\Model\TemplateFactory;
 use Magento\Framework\App\Area;
 use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Framework\App\RequestInterface;
-use Magento\Framework\App\TemplateTypesInterface;
 use Magento\Framework\Filesystem;
 use Magento\Framework\Filter\FilterManager;
 use Magento\Framework\Filter\Template;
+use Magento\Framework\Mail\TemplateTypesInterface;
 use Magento\Framework\Model\Context;
 use Magento\Framework\Registry;
 use Magento\Framework\Url;
