+++ app/code/Magento/Newsletter/view/adminhtml/templates/template/edit.phtml
@@ -4,7 +4,7 @@
  * See COPYING.txt for license details.
  */
 
-use Magento\Framework\App\TemplateTypesInterface;
+use Magento\Framework\Mail\TemplateTypesInterface;
 
 /* @var $block \Magento\Newsletter\Block\Adminhtml\Template\Edit */
 /** @var \Magento\Framework\View\Helper\SecureHtmlRenderer $secureRenderer */
